+++ finrl/README.md
@@ -1,7 +1,7 @@
 This folder has three subfolders:
 + applications: trading tasks,
 + agents: DRL algorithms, from ElegantRL, RLlib, or Stable Baselines 3 (SB3). Users can plug in any DRL lib and play.
-+ finrl_meta: market environments, we merge the stable ones from the active [FinRL-Meta repo](https://github.com/AI4Finance-Foundation/FinRL-Meta).
++ meta: market environments, we merge the stable ones from the active [FinRL-Meta repo](https://github.com/AI4Finance-Foundation/FinRL-Meta).
 
 Then, we employ a train-test-trade pipeline by three files: train.py, test.py, and trade.py.
 
@@ -17,7 +17,7 @@ FinRL
 │   	├── elegantrl
 │   	├── rllib
 │   	└── stablebaseline3
-│   ├── finrl_meta
+│   ├── meta
 │   	├── data_processors
 │   	├── env_cryptocurrency_trading
 │   	├── env_portfolio_allocation
