+++ tests/environments/test_cash_penalty.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from finrl.finrl_meta.env_stock_trading.env_stocktrading_cashpenalty import (
+from finrl.meta.env_stock_trading.env_stocktrading_cashpenalty import (
     StockTradingEnvCashpenalty,
 )
-from finrl.finrl_meta.preprocessor.yahoodownloader import YahooDownloader
+from finrl.meta.preprocessor.yahoodownloader import YahooDownloader
 
 
 @pytest.fixture(scope="session")
